implemented requests manager class, and updated reconciler implementation. Also, created index.
