+++ core/task-executor/lib/reconcile/managers/index.js
@@ -1,0 +1,9 @@
+const WorkersManager = require('./workers');
+const requestsManager = require('./requests');
+const jobsManager = require('./jobs');
+
+module.exports = {
+    WorkersManager,
+    requestsManager,
+    jobsManager
+};
